% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYTH.R
\name{MYTH}
\alias{MYTH}
\title{MYTH}
\usage{
MYTH(input_data_met)
}
\arguments{
\item{input_data_met}{A dataframe or matrix containing gene methylation profiles (beta-values) in tumors, in which the row name is gene symbol or Entrez gene ID and the column name is tumor sample ID.}
}
\value{
A dataframe with 2 columns:
\item{Sample}{Tumor samples to be calculated.}
\item{MYTH_score}{The MYTH score of each sample.}
}
\description{
\code{MYTH} A novel algorithm to score Methylation-Yielding Tumor Heterogeneity.
}
\details{
This function calculates the ITH score in a bulk tumor based on the variations of gene methylation levels.
}
\examples{
path <- system.file("extdata", "example_MYTH.txt", package = "MYTH", mustWork = TRUE)
input_data_met <- read.table(path, stringsAsFactors = FALSE, sep = "\\t", header = TRUE, quote = "", row.names = 1)
MYTH(input_data_met)
}
\author{
Lin Li <cpu_lilin@126.com>, Xiaosheng Wang <xiaosheng.wang@cpu.edu.cn>, Qian Liu <qian.liu@stu.cpu.edu.cn>
}
